game/requirements.txt
firebase-admin==3.*
flask==2.*
google-cloud-pubsub==2.*
slackclient==1.*
